fix boo#898861
YaST2 control center (Qt) displays a duplicate entry for Services manager. It is because there are actually two .desktop files for this module, for backward compatibility, but only one should be displayed. The other one has a Hidden flag and should be ignored.

This change makes YaST2 control center ignore desktop file with Hidden flag activated.